% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-polars.R
\name{install_polars}
\alias{install_polars}
\title{Helper function to install \code{polars} package}
\usage{
install_polars()
}
\description{
Helper function to install \code{polars} package from
\href{https://community.r-multiverse.org}{\code{r-multiverse}}.
}
