% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lst_scripts}
\alias{lst_scripts}
\title{Standalone Script}
\format{
An object of class \code{list} of length 18.
}
\usage{
lst_scripts
}
\description{
Named list where the values are raw URL locations of standalone scripts
}
\keyword{datasets}
