% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-namespace.R
\name{new_series}
\alias{new_series}
\alias{Series}
\title{Create Series object}
\usage{
new_series(value, name = "", ...)
}
\arguments{
\item{value}{Series value.}

\item{name}{Series names}

\item{...}{Additional attributes to be added.}
}
\description{
Create Series object
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Get the names of the Series}

\item{\code{len}}{Get the length of the Series}

\item{\code{dtype}}{Get the data type of the Series}

\item{\code{struct}}{Use struct methods}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeriesNamespace-slice}{\code{Series$slice()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeriesNamespace-slice"></a>}}
\if{latex}{\out{\hypertarget{method-SeriesNamespace-slice}{}}}
\subsection{Method \code{slice()}}{
Subset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Series$slice(offset, length = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{offset}}{Start 0-based index.}

\item{\code{length}}{The length of the subset}
}
\if{html}{\out{</div>}}
}
}
}
