% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{standalone-package}
\alias{standalone}
\alias{standalone-package}
\title{standalone: Standalone Scripts for package delopment}
\description{
Provide standalone scripts for `usethis::use_standalone()`.
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
