% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_standalone_script.R
\name{copy_standalone_script}
\alias{copy_standalone_script}
\title{Copy Standalone Script}
\usage{
copy_standalone_script(script, destdir = "./R")
}
\arguments{
\item{script}{(\code{string})\cr
named of the script to copy. The names follow the pattern \verb{<pkgname>-<script_name>}.
Must be one of 'standalone-checks', 'standalone-stringr', 'standalone-forcats', 'rlang-purrr'.}

\item{destdir}{directory where the script will be saved. Default is \code{"./R"}}
}
\description{
Copy a standalone script from within this package or from the rlang
package into your local package.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
copy_standalone_script("standalone-checks")
\dontshow{\}) # examplesIf}
}
