% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-package.R
\docType{package}
\name{standalone-package}
\alias{standalone}
\alias{standalone-package}
\title{standalone: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.danieldsjoberg.com/standalone/}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

}
\keyword{internal}
