% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-package.R
\docType{package}
\name{standalone-package}
\alias{standalone}
\alias{standalone-package}
\title{standalone: Package Housing Helper Functions}
\description{
Standalone shims for functions from packages such as 'forcats', 'stringr', etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/standalone/}
  \item \url{https://insightsengineering.github.io/standalone}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

}
\keyword{internal}
